Phase 7: Migrate :core:designsystem to KMP + Compose Multiplatform
Convert build to kotlin.multiplatform + compose.multiplatform.
Move all Compose components and theme to commonMain; move SaCIcons
and UserInfoPanel (use Android R.drawable/R.string) to androidMain.
Migrate font to commonMain/composeResources/; update Type.kt to use
Res.font.press_start_2p via CMP resources API.
Remove android.os.Build reference from Theme.kt (dynamic color unused).
Keep src/main/res/ Android resources for backward compat with
Android-only feature modules until Phase 9.
