+++ core/designsystem/src/commonMain/kotlin/com/maxot/seekandcatch/core/designsystem/RoundedTriangleShape.kt
