+++ core/designsystem/src/commonMain/kotlin/com/maxot/seekandcatch/core/designsystem/component/PixelFigureDrawer.kt
